Use EPISODE query to estimate the airing season of series
This implements the EPISODE query, per-aid caching of ep1 air date
response, and adds three new template fields (season, season_quarter,
season_year). season_year is essentially equivalent to anime_year but,
if the series starts airing in december, points to the next year instead.

The quarters / seasons used are roughly equivalent to the ones used in
http://cal.syoboi.jp/quarter/, but there are some differences to cope
with series that start a few days or weeks before what would've been
the season proper.